refactor: centralize id validation utils
