+++ src/utils/idValidation.js
@@ -1,17 +1,14 @@
-// src/entities/utils/idValidation.js
+// src/utils/idValidation.js
 
-import {
-  assertValidId,
-  assertNonBlankString,
-} from '../../utils/dependencyUtils.js';
-import { InvalidArgumentError } from '../../errors/invalidArgumentError.js';
+import { assertValidId, assertNonBlankString } from './dependencyUtils.js';
+import { InvalidArgumentError } from '../errors/invalidArgumentError.js';
 
 /**
  * Determine if a given value is a valid non-blank string ID.
  *
  * @param {any} id - Value to validate.
  * @param {string} context - Error context for logging.
- * @param {import('../../interfaces/coreServices.js').ILogger} logger - Logger instance.
+ * @param {import('../interfaces/coreServices.js').ILogger} logger - Logger instance.
  * @returns {boolean} True if the ID is valid, otherwise false.
  */
 export function isValidId(id, context, logger) {
@@ -28,7 +25,7 @@ export function isValidId(id, context, logger) {
  *
  * @param {string} instanceId - Entity instance ID to validate.
  * @param {string} componentTypeId - Component type ID to validate.
- * @param {import('../../interfaces/coreServices.js').ILogger} logger - Logger instance.
+ * @param {import('../interfaces/coreServices.js').ILogger} logger - Logger instance.
  * @param {string} [context] - Error context.
  * @throws {InvalidArgumentError} If either ID is invalid.
  * @returns {void}
